+++ libraries/dagster-sftp/dagster_sftp/__init__.py
@@ -6,6 +6,6 @@ from dagster_sftp.resource import (
     SFTPFileInfoConfig as SFTPFileInfoConfig,
 )
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 
 DagsterLibraryRegistry.register("dagster-sftp", __version__, is_dagster_package=False)