+++ libraries/dagster-sftp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.2]
+
+- Correctly pass `follow_symlinks` into `get_file_info()`
+- Add functions for validating authentication, `is_authenticated()` and `test_connection()`.
+- Add support for private keys as a path (`private_key_path`) or string (`private_key_content`).
+
 ## [0.0.1]
 
 - Initial Release, 2025-11-10